cleanup VideoFrame
