--- Tribler/Video/VideoFrame.py
@@ -1,62 +1,0 @@
-# Written by Fabian van der Werf and Arno Bakker
-# see LICENSE.txt for license information
-
-import wx
-import logging
-
-from Tribler.Video.defs import MEDIASTATE_PLAYING
-
-
-class VideoBaseFrame:
-
-    def show_videoframe(self):
-        pass
-
-    def hide_videoframe(self):
-        pass
-
-    def get_videopanel(self):
-        pass
-
-    def delete_videopanel(self):
-        pass
-
-    def OnCloseWindow(self, event=None):
-        pass
-
-    def get_window(self):
-        pass
-
-    def ShowLoading(self):
-        pass
-
-    def Stop(self):
-        pass
-
-    def Destroy(self):
-        pass
-
-
-# See Tribler/Main/vwxGUI/MainVideoFrame.py for Tribler 5.0
-# See Tribler/Player/PlayerVideoFrame.py for the SwarmPlayer / 4.5
-
-
-class DelayTimer(wx.Timer):
-
-    """ vlc.MediaCtrl needs some time to stop after we give it a stop command.
-        Wait until it is and then tell it to play the new item
-    """
-    def __init__(self, embedplay):
-        self._logger = logging.getLogger(self.__class__.__name__)
-
-        wx.Timer.__init__(self)
-        self.embedplay = embedplay
-        self.Start(100)
-
-    def Notify(self):
-        if self.embedplay.GetState() != MEDIASTATE_PLAYING:
-            self._logger.debug("embedplay: VLC has stopped playing previous video, starting it on new")
-            self.Stop()
-            self.embedplay.Play()
-        else:
-            self._logger.debug("embedplay: VLC is still playing old video")
